requirements.txt
pandas~=2.1.3
scikit-learn~=1.3.2
numpy~=1.24.3
Pillow~=10.1.0
keras~=2.15.0
matplotlib~=3.8.1
tensorflow~=2.15.0